refactored block manager so that loaded block index is separate from pending block index expanded block.Info to include additional details, removed packBlockInfo removed repackaging support for now, since it's greatly complicating the implementation

+++ cli/command_block_list.go
@@ -30,7 +30,7 @@ func runListBlocksAction(ctx context.Context, rep *repo.Repository) error {
 	var totalSize int64
 	uniquePacks := map[block.PhysicalBlockID]bool{}
 	for _, b := range blocks {
-		totalSize += b.Length
+		totalSize += int64(b.Length)
 		count++
 		if b.PackBlockID != "" {
 			uniquePacks[b.PackBlockID] = true
